refactor(core): expose root observability module
